% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jumps_ages.r
\name{calculate_jumps_ages}
\alias{calculate_jumps_ages}
\title{calculate_jumps_ages}
\usage{
calculate_jumps_ages(database, id_field, rounded_ages_field,
  diameter_class_field)
}
\arguments{
\item{database}{data.frame data.table}

\item{id_field}{string name of column with unique identifier of each tree}

\item{rounded_ages_field}{string name of column with rounded ages}

\item{diameter_class_field}{string name of column with diametric classes}
}
\description{
This function generates the jumps per diameter class among the ages in the database
}

