% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.integer.R
\name{check.integer}
\alias{check.integer}
\title{Ckeck Integer}
\usage{
check.integer(x)
}
\arguments{
\item{x}{any variable}
}
\value{
TRUE if "x" is integer, FALSE if "x" not is interger
}
\description{
checks if a variable is integer
}
\examples{
x = 5
check.integer(x)
}
