% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ages.r
\name{check_ages}
\alias{check_ages}
\title{check_ages}
\usage{
check_ages(data_base, rounded_age1, rounded_age2, months = FALSE)
}
\arguments{
\item{data_base}{data.frame data.table}

\item{rounded_age1}{string name of column rounde age one}

\item{rounded_age2}{string name of column rounde age two}

\item{months}{TRUE for age in months or FALSE for age in years}
}
\description{
This analysis verifies age differences on a paired basis, if the rounded ages are in months the check is if the difference is 12 months, if it is in year the consecutive ages should only present difference of 1 year, doubts about how to pair your base consult The Fgmutils package
}
