% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clone_different_parcel.r
\name{check_clones_different_parcel}
\alias{check_clones_different_parcel}
\title{check_clones_different_pacel}
\usage{
check_clones_different_parcel(database, parcel_name, clone_name,
  variables_to_group)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{parcel_name}{string name of the field containing the parcels}

\item{clone_name}{string name of the field containing the clones}

\item{variables_to_group}{string(s) variable (s) that you want to group the result of the analysis}
}
\description{
This function checks if the clones of a tree have different plots
}

