% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dead_state.r
\name{check_dead_state}
\alias{check_dead_state}
\title{check_dead_state}
\usage{
check_dead_state(data_base, state, measurement_variables)
}
\arguments{
\item{data_base}{data.frame data.table or any database}

\item{state}{string field name representing state column in database}

\item{measurement_variables}{string vector that contains a set of measurement variables to be analyzed, this variables are names of columns in database}
}
\description{
This function checks if the base state field is equal to dead (M) and there is some kind of measurement
}

