% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_diametric_increase.r
\name{check_diametric_increase}
\alias{check_diametric_increase}
\title{check_diametric_increase}
\usage{
check_diametric_increase(database, id_field, plot_field, plot_analyzed,
  ages_field, rounded_ages_field, diameter_classes_field, database_nhaclasses,
  plot_field_database_nhaclassses, diameter_classes_field_database_nhaclasses,
  rounded_ages_field_database_nhaclasses, nha_classes_database_nhaclasses)
}
\arguments{
\item{database, }{data.table, data.frame or any database}

\item{id_field, }{string with the name of column containing the unique identifier of each tree}

\item{plot_field, }{string with the name of column containing the plots of database}

\item{plot_analyzed, }{the number of plot to be analyzed, example: 356}

\item{ages_field, }{string with the name of column containing the ages of each tree}

\item{rounded_ages_field, }{string with the name of column containing the rounded ages of each tree, example: 12, 36}

\item{diameter_classes_field, }{string with the name of column containing the diameter class of each tree}

\item{database_nhaclasses, }{data.table, data.frame, or any database containg the NHa Class of database used in this analysis. To obtain this database you simply have to submit the original base the generate_number_hectare_class function contained in this package}

\item{plot_field_database_nhaclassses, }{string with the name of column containing the plots of database resulting from the generate_number_hectare_class function}

\item{diameter_classes_field_database_nhaclasses, }{string with the name of column containing the diameter class of each tree of database resulting from the generate_number_hectare_class function}

\item{rounded_ages_field_database_nhaclasses, }{string with the name of column containing the rounded ages of each tree of database resulting from the generate_number_hectare_class function}

\item{nha_classes_database_nhaclasses, }{string with the name of column containing the NHa Class of database resulting from the generate_number_hectare_class function}
}
\description{
This function is an implementation of the process proposed by Demolinari (2006) in his master's thesis that is the diametric increase, given a plot it is possible to follow all the growth of the trees contained in the same at all ages present in the plot, this is done through Of colored color graphics where each color represents a class of diameter and a meditation that as an advance for each age is generated a new graphic that contains as changes represented by cores or a quantity of trees that migrate of class, this can have an identification Of inconsistent behaviors within the plot.
}

