% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_ages.r
\name{check_existing_ages}
\alias{check_existing_ages}
\title{check_existing_ages}
\usage{
check_existing_ages(database, ages_name, ages_to_check)
}
\arguments{
\item{database}{data.frame data.table or any database}

\item{ages_name}{string name of the column representing ages}

\item{ages_to_check}{string name/vector of the column (s) representing ages to be checked}
}
\description{
This function checks if a given set of ages exists in a database column
}

