% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_place.r
\name{check_existing_place}
\alias{check_existing_place}
\title{check_existing_place}
\usage{
check_existing_place(database, place_name, places_to_check)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{place_name}{string name of the column representing site or place}

\item{places_to_check}{value(s) to be checked, example: c(12,21,33)}
}
\description{
This function checks whether a particular set of sites or locations exists in a database column
}

