% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_plot.r
\name{check_existing_plots}
\alias{check_existing_plots}
\title{check_existing_plots}
\usage{
check_existing_plots(database, plots_name, plots_to_check)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{plots_name}{string column name representing parcels in the base}

\item{plots_to_check}{value(s) to be checked, example: c(356,122)}
}
\description{
This function checks if a particular set of parcels exists in a database column
}

