% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_measurement_ages.r
\name{check_measurement_ages}
\alias{check_measurement_ages}
\title{check_measurement_ages}
\usage{
check_measurement_ages(data_base, measurement_variable1, measurement_variable2)
}
\arguments{
\item{data_base}{data.frame, data.table or any database}

\item{measurement_variable1}{string field containing the measurement variables at age 1}

\item{measurement_variable2}{string field containing the measurement variables at age 2}
}
\description{
This function verifies if measurement variables have records of type DAP2 <DAP1, HT2 <HT1 in consecutive ages i + 1 and i tt is necessary that the base is already paired to perform such analysis, to know more about pairing consult the Fgmutils package
}

