% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_measurements_state.r
\name{check_measurements_state}
\alias{check_measurements_state}
\title{check_measurements_state}
\usage{
check_measurements_state(data_base, measurement_variables, state)
}
\arguments{
\item{data_base}{data.frame, data.table or any database}

\item{measurement_variables}{set of variables to be analyzed, this set can be a vector of string with names of colunms}

\item{state}{string name of the field that represents the state in database}
}
\description{
This function checks if there is a measurement variable with value equal to 0 and if the respective states are different from M, F, A
}

