% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parcel_different_spacing.r
\name{check_parcel_different_spacing}
\alias{check_parcel_different_spacing}
\title{check_parcel_different_spacing}
\usage{
check_parcel_different_spacing(database, parcel_name, spacing_age1,
  spacing_age2, variables_to_group)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{parcel_name}{string containing the field name parcels in database}

\item{spacing_age1}{string containing the name of the field spacing in the first age}

\item{spacing_age2}{string containing the name of the field spacing in the second age}

\item{variables_to_group}{variable (s) that you want to group the result of the analysis, this can be a vector os strings or strign name to group}
}
\description{
This function checks for partitions with different spacing at i and i + 1 ages, it is necessary that the base be paired including the field representing the spacing, doubts about how to pair its base see the Fgmutils package
}

