% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_size_age.r
\name{check_size_age_parcel}
\alias{check_size_age_parcel}
\title{check_size_age_parcel}
\usage{
check_size_age_parcel(database, age_name)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{age_name}{string containing the name of the column that represents age}
}
\description{
This function checks if the age field is more than one age, returning TRUE to for yes and FALSE for no
}

