% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_undefined_spacing.r
\name{check_undefined_spacing}
\alias{check_undefined_spacing}
\title{check_undefined_spacing}
\usage{
check_undefined_spacing(data_base, spacings)
}
\arguments{
\item{data_base}{data.frame, data.table or any database}

\item{spacings}{string vector containing the name of the variable (s) than represent spacings in database}
}
\description{
This function checks if there is any record with undefined spacing (0 or NA)
}

