% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variables.r
\name{check_variables}
\alias{check_variables}
\title{check_variables}
\usage{
check_variables(database, variables)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{variables}{vector of strings with names of columns}
}
\value{
TRUE for all variables in database, or FALSE for variables not present in columns
}
\description{
This function checks if the entered column exists within the base
}
\examples{
test <- data.frame("tree","diametrer","N")
check_variables(test,c("tree","diameter"))
}

