% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_zero_measurement.r
\name{check_zero_measurement}
\alias{check_zero_measurement}
\title{check_zero_measurement}
\usage{
check_zero_measurement(data_base, measurement_variables, state_name,
  states_to_check)
}
\arguments{
\item{data_base}{data.frame, data.table or any database}

\item{measurement_variables}{string vector containing name of the field(s) it represents measurement variable(s) to be analyzed}

\item{state_name}{string vector containing the name of the variable than represents state in database}

\item{states_to_check}{string vector containing the name of the the states to be checked, the user can inform this names in a string vector like ("F","N")}
}
\description{
This analysis verifies which measurement variables have values equal to 0 and then checks if there are variables in the states that the user reported
}
