% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_probability.r
\name{correct_probability}
\alias{correct_probability}
\title{correct_probability}
\usage{
correct_probability(database, plot_field, rounded_age_field, probability_field,
  nha_class_field)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{plot_field}{string name of column with plots of database}

\item{rounded_age_field}{string name of column with rounded ages, examples: 12,24,36,48}

\item{probability_field}{string name of column with probabilitys}

\item{nha_class_field}{string name of column with NHa (surviving tree classes per hectare) classes,}
}
\description{
This function will adjust the probability after generation of surviving tree classes per hectare
}

