% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_age.r
\name{find_missing_age}
\alias{find_missing_age}
\title{find_missing_age}
\usage{
find_missing_age(database, age_name, ages_to_check)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{age_name}{string that contains the field name that represents age in database}

\item{ages_to_check}{vector containing the values of ages to be checked like c(12,23,48)}
}
\description{
This function identifies the missing age values in the database and notifies them to the user.
}
