% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_place.r
\name{find_missing_place}
\alias{find_missing_place}
\title{find_missing_place}
\usage{
find_missing_place(database, place_name, places_to_check)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{place_name}{string that contains the field name representing site or place in database}

\item{places_to_check}{vector containing the values of places/sites to be checked like c(21,33,48)}
}
\description{
This function identifies values of sites or locations in the database and notifies them to the user
}
