% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_variable.r
\name{find_missing_variable}
\alias{find_missing_variable}
\title{find_missing_variable}
\usage{
find_missing_variable(data_base, variables)
}
\arguments{
\item{data_base}{data.frame, data.table or any database}

\item{variables}{vector string that contains the name(s) of columns to be checked in database}
}
\description{
This function identifies non-existent column names in the database and informs the user
}
