% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_colors_diameter_class.r
\name{generate_colors_diameter_class}
\alias{generate_colors_diameter_class}
\title{generate_colors_diameter_class}
\usage{
generate_colors_diameter_class(database, diameter_classe_name)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{diameter_classe_name}{string with the name of field (column) containing the diameter classes}
}
\description{
This function generates a new random color for each diameter class in the base
}

