% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_diameter_classes.r
\name{generate_diameter_classes}
\alias{generate_diameter_classes}
\title{generate_diameter_classes}
\usage{
generate_diameter_classes(database, diameter_names, amplitude,
  name_of_diameter_class)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{diameter_names}{string with name of the field that contains the diameters of database}

\item{amplitude}{desired amplitude for class creation, example: 1,2,4,6,7}

\item{name_of_diameter_class}{string with name you want for the field class of diameter}
}
\description{
This function identifies non-existent column names in the database and informs the user
}

