% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_diameter_class.r
\name{generate_initial_diameter_class}
\alias{generate_initial_diameter_class}
\title{generate_initial_diameter_class}
\usage{
generate_initial_diameter_class(database, plot_name, age_name)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{plot_name}{string with the name of field representing plots in database}

\item{age_name}{string with the name of field representing rounded age}
}
\description{
This function generates the initial class field to aid in the process of diametric increasing
}
