% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_limiar.r
\name{generate_limiar}
\alias{generate_limiar}
\title{generate_limiar}
\usage{
generate_limiar(database, field, max)
}
\arguments{
\item{database, }{data.table, data.frame or any database}

\item{field, }{string name with column you want to work on}

\item{max, }{number with the maximum value you want to establish the threshold}
}
\description{
This function generates a threshold for certain values based on the field entered and a maximum value for setting
}

