% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_new_color.r
\name{generate_new_color}
\alias{generate_new_color}
\title{generate_new_color}
\usage{
generate_new_color(colors)
}
\arguments{
\item{colors}{vector of strings containing existing colors, exemple: c("#6140bc" "#e75bf7" "#d15102" "#6a0b9e" "#e8ad4e")}
}
\description{
This function generates a new random color without repeating the ones that were entered in the last field as parameter
}

