% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_number_hectare.r
\name{generate_number_hectare}
\alias{generate_number_hectare}
\title{generate_number_hectare}
\usage{
generate_number_hectare(database, area_name, n_name, nha_name = "NHa")
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{area_name}{string with the name of field containing area in database}

\item{n_name}{string with the name of field containing numbers of trees in database}

\item{nha_name}{string with name you want for the field number of trees per hectare}
}
\description{
This function generates the NHa, field that represents the number of surviving trees per hectare
}

