% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_number_hectare_class.r
\name{generate_number_hectare_class}
\alias{generate_number_hectare_class}
\title{generate_number_hectare_class}
\usage{
generate_number_hectare_class(database, nha_field, n_field, area_field,
  plot_field, rounded_age_field, age_field, diameter_class_field, state_field,
  dap_field, amplitude = 2)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{nha_field, }{name of the column containing the number of trees surviving per hectare}

\item{n_field, }{name of the column containing the number of trees in a plot, or diameter class}

\item{area_field, }{name of the column containing the area}

\item{plot_field, }{name of the column containing the plots}

\item{rounded_age_field, }{name of the column containing the rounded ages, example: 12,24,36}

\item{age_field, }{name of the column containing the  ages}

\item{diameter_class_field, }{name of the column containing the diameter classes}

\item{state_field, }{name of the column containing the states of trees}

\item{dap_field, }{name of the column containing the diameter (DAP) of trees}

\item{amplitude, }{integer integer number with the amplitude of diameter classes, example: 1, 2, 3}
}
\description{
This function generates the NHa classes column which represents the classes with the number of surviving trees per hectare
}

