% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ages.r
\name{get_ages}
\alias{get_ages}
\title{get_ages}
\usage{
get_ages(database, age_name, age_values)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{age_name}{string with the name of field (column) containing the ages}

\item{age_values}{vector with the age values you want to assemble string to made query, example: c(12,24,36)}
}
\description{
This function concatenates age values in a string for a query and returns the same
}
