% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max.r
\name{get_max}
\alias{get_max}
\title{get_max}
\usage{
get_max(database, variables)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{variables}{string vector with name(s) of the column (s) you want to know the maximum value}
}
\description{
This function returns the maximum value of one or more fields of measurement variables
}
