% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min.r
\name{get_min}
\alias{get_min}
\title{get_min}
\usage{
get_min(database, variables)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{variables}{string vector with name(s) of the column (s) you want to know the minimum value}
}
\description{
This function returns the minimum value of one or more fields of measurement variables
}
