% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_place.r
\name{get_place}
\alias{get_place}
\title{get_place}
\usage{
get_place(database, place_name, place_value)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{place_name}{string with the name of the column that represents the place}

\item{place_value}{vector with values of that you want to filter the sites/places of the database}
}
\description{
This function returns a database from a particular site or location present in the original database
}
