% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColors.r
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors(database, diameter_classe_name)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{diameter_classe_name}{string with the name of field (column) containing the diameter classes}
}
\description{
This function generates a new random color for each diameter class in the base
}

