% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount_list.r
\name{mount_list}
\alias{mount_list}
\title{mount_list}
\usage{
mount_list(database, fieldColorName = "color", fieldQuantityName = "n",
  diameter_classes_name = "classedediametro")
}
\arguments{
\item{database, }{data.table, data.frame or any database}

\item{fieldColorName, }{string with the name you want for the color column}

\item{fieldQuantityName, }{string name of column with number of trees}

\item{diameter_classes_name, }{string name of column with the diameter classes}
}
\description{
This function serves as a help for the check_diametric_increase function, it generates lists of values to be plotted as columns
}

