% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount_query.r
\name{mount_query}
\alias{mount_query}
\title{mount_query}
\usage{
mount_query(database, select_names, group_names, option)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{select_names}{string vector with the name(s) of the column(s) you want to include in the selection}

\item{group_names}{string vector with the name(s) of the column(s) you want to group the results}

\item{option}{options to make the query, can be 1,2,3 each one for one use in the analysis functions}
}
\description{
This auxiliary function checks that need to group fields of certain measurements
}

