% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cols.r
\name{plot_cols}
\alias{plot_cols}
\title{plot_cols}
\usage{
plot_cols(list, name, my_title, ceil = NaN)
}
\arguments{
\item{list, }{list of columns generated by function mount_list with values and colors to be ploted}

\item{name, }{string name with the name you want for the file png chart}

\item{my_title, }{string with the name you want for the title of chart}

\item{ceil, }{the maximum value to axis, this value will be applied in NHa or N}
}
\description{
This function serves as a help for the check_diametric_increase function, it generates a chart based on the lists created by the mount_list function
}

