% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_columns_crescent.r
\name{sort_columns_crescent}
\alias{sort_columns_crescent}
\title{sort_columns_crescent}
\usage{
sort_columns_crescent(database, column)
}
\arguments{
\item{database}{data.frame, data.table or any database}

\item{column}{string with the name of the column you want sort the database}
}
\description{
Sorts the database incrementally based on the selected column
}

